Add interrupt for external slide detection circuit
This doesn't actually do much yet since the microprocessor only sleeps
for at most 1 ms until the next interrupt for millis(), and it's using
the highest power sleep mode (idle).

This will actually start to be useful when using a crystal for the clock
instead of the Arduino library's (interrupt-based) millis().