Prevent double-firing of Timer2 overflow interrupt
This can happen if sleep is entered too quickly after the interrupt
occurs. As the datasheet recommends, write a value to one of the Timer2
registers and wait for the asynchronous register to clear to ensure a
cycle has passed and it is safe to enter a low power mode.